requirements_pinned.txt
# Runtime dependencies

cycler==0.12.1
numpy==1.26.4
matplotlib==3.9.2
pyglotaran==0.7.3
tabulate==0.9.0
xarray==2024.7.0
ruamel.yaml==0.18.6
docstring_parser==0.16
pydantic==2.7.2
